Update version check to close the node process again
This updates the version check to use the block syntax for IO#popen.
Doing so will close the IO object and also terminate the node subprocess
again. The result of the block is returned, so the behavior of the
version check is exactly the same as before, it now just cleans up after
itself.